add approve procee

+++ routes/dashboard.php
@@ -43,12 +43,14 @@ Route::middleware('jwt')->group(function () {
     });
 
     Route::group(['prefix' => 'partner'], function () {
-        Route::get('/generate-password', 'PartnerController@generatePassword');
-        Route::post('/', 'PartnerController@store');
-        Route::get('/', 'PartnerController@index');
-        Route::get('/{id}', 'PartnerController@show');
-        Route::put('/{id}', 'PartnerController@update');
-        Route::delete('/{id}', 'PartnerController@destroy');
+        Route::get('/generate-password', 'PartnerController@generatePassword')->permission(PermissionEnum::PARTNER_GENERATE_PASSWORD->value);
+        Route::post('/', 'PartnerController@store')->permission(PermissionEnum::PARTNER_STORE->value);
+        Route::get('/', 'PartnerController@index')->permission(PermissionEnum::PARTNER_INDEX->value);
+        Route::get('/{id}', 'PartnerController@show')->permission(PermissionEnum::PARTNER_SHOW->value);
+        Route::put('/{id}', 'PartnerController@update')->permission(PermissionEnum::PARTNER_UPDATE->value);
+        Route::delete('/{id}', 'PartnerController@destroy')->permission(PermissionEnum::PARTNER_DESTROY->value);
+        Route::post('/{id}/approve', 'PartnerController@approveAccount')->permission(PermissionEnum::PARTNER_APPROVE_ACCOUNT->value);
+        Route::post('/{id}/approve-kyc', 'PartnerController@approveKYC')->permission(PermissionEnum::PARTNER_APPROVE_KYC->value);
     });
 
     Route::group(['prefix' => 'email-content'], function () {
